Design Matrix for a Model

This is a helper function to form the design matrix for an LNM regression starting from a fitted model's formula object. It is an analog of model.matrix for the multiresponse setting.

prepare_newdata(fit, newdata = NULL)

Arguments

fit

An object of class lnm whose estimate slot contains the rstan fitted logistic normal multinomial model.

newdata

A data.frame containing variables in the formula definition of the fit, but which hasn't been converted into the matrix format needed for internal prediction.

Value

A matrix containing the design matrix that can be multiplied with the fitted Beta parameter to get fitted compositions.

Examples

example_data <- lnm_data(N = 10, K = 5)
xy <- dplyr::bind_cols(example_data[c("X", "y")])
fit <- lnm(
    starts_with("y") ~ starts_with("x"), xy,
    iter = 5, output_samples = 5
)
#> Chain 1: ------------------------------------------------------------
#> Chain 1: EXPERIMENTAL ALGORITHM:
#> Chain 1:   This procedure has not been thoroughly tested and may be unstable
#> Chain 1:   or buggy. The interface is subject to change.
#> Chain 1: ------------------------------------------------------------
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Gradient evaluation took 3.6e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.36 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Begin eta adaptation.
#> Chain 1: Iteration:   1 / 250 [  0%]  (Adaptation)
#> Chain 1: Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Chain 1: Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Chain 1: Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Chain 1: Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Chain 1: Success! Found best value [eta = 1] earlier than expected.
#> Chain 1: 
#> Chain 1: Begin stochastic gradient ascent.
#> Chain 1:   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#> Chain 1: Informational Message: The maximum number of iterations is reached! The algorithm may not have converged.
#> Chain 1: This variational approximation is not guaranteed to be meaningful.
#> Chain 1: 
#> Chain 1: Drawing a sample of size 5 from the approximate posterior... 
#> Chain 1: COMPLETED.
#> Warning: Pareto k diagnostic value is Inf. Resampling is disabled. Decreasing tol_rel_obj may help if variational algorithm has terminated prematurely. Otherwise consider using sampling instead.
prepare_newdata(fit, example_data[["X"]])
#>               x1          x2        x3         x4         x5
#>  [1,] 0.93834241 0.198228124 0.2051782 0.30085451 0.03613237
#>  [2,] 0.84647114 0.357611314 0.6630767 0.26350152 0.02015370
#>  [3,] 0.57114358 0.181333204 0.4637440 0.30504613 0.37278110
#>  [4,] 0.67909341 0.561144860 0.3603482 0.87495827 0.22566540
#>  [5,] 0.08932224 0.659083010 0.6893664 0.71843722 0.49191810
#>  [6,] 0.78537413 0.660253409 0.2588608 0.40691969 0.44664516
#>  [7,] 0.22695745 0.002400788 0.8512050 0.01935947 0.69020490
#>  [8,] 0.44817847 0.993446034 0.4800047 0.05018036 0.84273998
#>  [9,] 0.16122764 0.627489231 0.5372664 0.70225103 0.37583367
#> [10,] 0.17611169 0.014630336 0.6616399 0.06637027 0.98488969